requirements.txt
matplotlib==3.5.1
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.2.1
sklearn_crfsuite==0.3.6
torch==2.1.0
TorchCRF==1.1.0
TorchCRF==1.1.0
transformers==4.30.2
